Pull out a single variable

This works like [[ for local data frames, and automatically collects before indexing for remote data tables.

# S3 method for FDF
pull(.data, var = -1)

pull(.data, var = -1)

Arguments

.data
A FDF.

var
A variable specified as:

a literal variable name

a positive integer, giving the position counting from the left

a negative integer, giving the position counting from the right.

The default returns the last column (on the assumption that's the column you've created most recently).

This argument is taken by expression and supports quasiquotation (you can unquote column names and column positions).

Examples

test <- FDF(datasets::iris)
pull(test, 1)

#>   [1] 5.1 4.9 4.7 4.6 5.0 5.4 4.6 5.0 4.4 4.9 5.4 4.8 4.8 4.3 5.8 5.7 5.4 5.1
#>  [19] 5.7 5.1 5.4 5.1 4.6 5.1 4.8 5.0 5.0 5.2 5.2 4.7 4.8 5.4 5.2 5.5 4.9 5.0
#>  [37] 5.5 4.9 4.4 5.1 5.0 4.5 4.4 5.0 5.1 4.8 5.1 4.6 5.3 5.0 7.0 6.4 6.9 5.5
#>  [55] 6.5 5.7 6.3 4.9 6.6 5.2 5.0 5.9 6.0 6.1 5.6 6.7 5.6 5.8 6.2 5.6 5.9 6.1
#>  [73] 6.3 6.1 6.4 6.6 6.8 6.7 6.0 5.7 5.5 5.5 5.8 6.0 5.4 6.0 6.7 6.3 5.6 5.5
#>  [91] 5.5 6.1 5.8 5.0 5.6 5.7 5.7 6.2 5.1 5.7 6.3 5.8 7.1 6.3 6.5 7.6 4.9 7.3
#> [109] 6.7 7.2 6.5 6.4 6.8 5.7 5.8 6.4 6.5 7.7 7.7 6.0 6.9 5.6 7.7 6.3 6.7 7.2
#> [127] 6.2 6.1 6.4 7.2 7.4 7.9 6.4 6.3 6.1 7.7 6.3 6.4 6.0 6.9 6.7 6.9 5.8 6.8
#> [145] 6.7 6.7 6.3 6.5 6.2 5.9

pull(test, 5)

#>   [1] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>   [6] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [11] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [16] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [21] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [26] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [31] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [36] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [41] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [46] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [51] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [56] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [61] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [66] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [71] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [76] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [81] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [86] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [91] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [96] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#> [101] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [106] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [111] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [116] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [121] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [126] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [131] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [136] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [141] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [146] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 

pull(test, Species)

#>   [1] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>   [6] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [11] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [16] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [21] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [26] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [31] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [36] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [41] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [46] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [51] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [56] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [61] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [66] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [71] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [76] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [81] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [86] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [91] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [96] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#> [101] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [106] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [111] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [116] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [121] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [126] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [131] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [136] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [141] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [146] "virginica"  "virginica"  "virginica"  "virginica"  "virginica"